Moved entities classes to own seperate classes
