+++ src/Chirp.Infrastructure/CheepRepository.cs
@@ -1,6 +1,7 @@
 using System.Globalization;
 using Chirp.Core;
 using Chirp.Core.DTOs;
+using Chirp.Infrastructure.Entities;
 
 namespace Chirp.Infrastructure;
