+++ src/Chirp.Infrastructure/Entities/Author.cs
@@ -1,0 +1,15 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Chirp.Infrastructure.Entities;
+
+public class Author
+{
+    [Key]
+    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+    public Guid AuthorId { get; set; }
+    
+    public required string Name { get; set; }
+    public required string Email { get; set; }
+    public IEnumerable<Cheep> Cheeps { get; set; } = new List<Cheep>();
+}
